+++ src/test/java/com/zipcodewilmington/danny_do_better_exercises/mathutilities/TestMultiplication.java
@@ -53,7 +53,7 @@ public class TestMultiplication {
         byte multiplier = 14;
         byte expectedByte = -32;
         // : When
-        byte actualByte = mathUtils.multiply(multiplicand, multiplier);
+        int actualByte = mathUtils.multiply(multiplicand, multiplier);
         // : Then
         assertEquals(expectedByte, actualByte);
     }
